2nd Commit
Tried to get the navbar to fix to top after scroll but was unable to do
that so I changed in the rinjani.css that it is just fixed to the top.
Adjusted images and deleted the ones I didn't use.
Also deleted what wasn't used from the rinjani theme in the .html and
.css files.